refactor to use VersionModifier struct
This renames NewVersion to ModifierType and wraps it in a
VersionModifier struct that can optionally include build
and pre-release data.

This also renames the metadata field to be build or
buid_metadata as appropriate. This matches the underlying
semver library better, as well as how *I* conceptualize it.

Additional tests across the config and main layers.