run_build option to fix issues with Cargo.lock
When running with the git tag option, the Cargo.lock file is not updated. This results in a situation where on the next `cargo build` the Carco.lock file will be updated with the new version. To get these to be updated in the same commit, we can run `cargo build` beforehand which will update anything in the Cargo.lock that's required. It also has a small side effect of ensuring that the Cargo.toml file has not been modified in such a way to break the build.

Adding this as an options makes this a backwards compatible change.